+++ libraries/package-deps-hash/src/index.ts
@@ -16,6 +16,8 @@
 export { getPackageDeps, getGitHashForFiles } from './getPackageDeps';
 export {
   type IFileDiffStatus,
+  type IDetailedRepoState,
+  getDetailedRepoStateAsync,
   getRepoChanges,
   getRepoRoot,
   getRepoStateAsync,