+++ libraries/rush-lib/src/logic/operations/OperationExecutionRecord.ts
@@ -24,15 +24,15 @@ import { OperationMetadataManager } from './OperationMetadataManager';
 import type { IPhase } from '../../api/CommandLineConfiguration';
 import type { RushConfigurationProject } from '../../api/RushConfigurationProject';
 import { CollatedTerminalProvider } from '../../utilities/CollatedTerminalProvider';
+import type { IOperationExecutionResult } from './IOperationExecutionResult';
+import type { IInputsSnapshot } from '../incremental/InputsSnapshot';
+import { RushConstants } from '../RushConstants';
+import type { IEnvironment } from '../../utilities/Utilities';
 import {
   getProjectLogFilePaths,
   type ILogFilePaths,
   initializeProjectLogFilesAsync
 } from './ProjectLogWritable';
-import type { IOperationExecutionResult } from './IOperationExecutionResult';
-import type { IInputsSnapshot } from '../incremental/InputsSnapshot';
-import { RushConstants } from '../RushConstants';
-import type { IEnvironment } from '../../utilities/Utilities';
 
 export interface IOperationExecutionRecordContext {
   streamCollator: StreamCollator;
@@ -381,21 +381,24 @@ export class OperationExecutionRecord implements IOperationRunnerContext, IOpera
         this.status = this.operation.enabled
           ? await this.runner.executeAsync(this)
           : this.runner.isNoOp
-            ? OperationStatus.NoOp
-            : OperationStatus.Skipped;
+          ? OperationStatus.NoOp
+          : OperationStatus.Skipped;
       }
+      // Make sure that the stopwatch is stopped before reporting the result, otherwise endTime is undefined.
+      this.stopwatch.stop();
       // Delegate global state reporting
       await onResult(this);
     } catch (error) {
       this.status = OperationStatus.Failure;
       this.error = error;
+      // Make sure that the stopwatch is stopped before reporting the result, otherwise endTime is undefined.
+      this.stopwatch.stop();
       // Delegate global state reporting
       await onResult(this);
     } finally {
       if (this.isTerminal) {
         this._collatedWriter?.close();
         this.stdioSummarizer.close();
-        this.stopwatch.stop();
       }
     }
   }