+++ heft-plugins/heft-sass-load-themed-styles-plugin/CHANGELOG.md
@@ -1,6 +1,16 @@
 # Change Log - @rushstack/heft-sass-load-themed-styles-plugin
 
-This log was last generated on Thu, 01 May 2025 15:11:33 GMT and should not be manually modified.
+This log was last generated on Thu, 15 May 2025 00:11:49 GMT and should not be manually modified.
+
+## 0.1.7
+Thu, 15 May 2025 00:11:49 GMT
+
+_Version update only_
+
+## 0.1.6
+Tue, 13 May 2025 02:09:20 GMT
+
+_Version update only_
 
 ## 0.1.5
 Thu, 01 May 2025 15:11:33 GMT