+++ libraries/package-deps-hash/CHANGELOG.md
@@ -1,6 +1,18 @@
 # Change Log - @rushstack/package-deps-hash
 
-This log was last generated on Thu, 01 May 2025 15:11:33 GMT and should not be manually modified.
+This log was last generated on Tue, 13 May 2025 02:09:20 GMT and should not be manually modified.
+
+## 4.4.1
+Tue, 13 May 2025 02:09:20 GMT
+
+_Version update only_
+
+## 4.4.0
+Thu, 08 May 2025 00:11:15 GMT
+
+### Minor changes
+
+- Add `getDetailedRepoState` API to expose `hasSubmodules` and `hasUncommittedChanges` in addition to the results returned by `getRepoState`.
 
 ## 4.3.24
 Thu, 01 May 2025 15:11:33 GMT